requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
pandas==2.3.1
numpy==2.3.1
psycopg2-binary==2.9.7
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0

# Vector Search & ML (Phase 2)
pgvector==0.2.4
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Background Tasks (Phase 2)
celery==5.3.2
redis==5.0.0

# Development & Testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0 